requirements.txt
apache-airflow~=2.3.4
apache-airflow-providers-docker~=3.1.0